Complete the /register endpoint
Write the logic for an endpoint that registers the new user.

Validate request using express validator for correct inputs.
Use bcryptjs to hash password for this new user before
saving to the database. When saved, create a new jwt token
for this user and save this token in their cookie.

Change the design for setting up server. Use node-built-in
http to create the main server and provide it the app instance
of express.

Test this new endpoint using Postman.