use react query to register user
In app.ts in backend folder, add the origin for the client.

Update the register function in requests.ts. The body is being
passed into the headers properties. This is causing an error
as the data is not getting passed into the request body.

Add the ButtonLoader component and check whether the user is
gettting registered and re-directed to the Home page.