Add another method to validate token for isLoggedIn
To let the client know whether the user is logged in or not,
create a varible is the context whose value will determine
upon the data fetched from "validate-token" endpoint.

Create a new method in request.ts that will make this
fetch request to the above endpoint. Use useQuery()
and use its return value to detemine the value for
isLoggedIn.