Add invalidation of query when the user registers or logs in
The validateToken query must be updated as soon as the
user registers or logs in. That's how the isLoggedIn value
in context can be updated. In Register and Login components,
make use of useQueryClient() hook and add this logic when
the registration or logging in is successfull.

Also, create another function that makes request to the
"logout" endpoint in request.ts.