Create the HeaderLogoutBtn
Create a log out button component that will encapsulate
all the logic of logging out a user.

Clicking on this button should use the logout function in
request.ts. It should also invalidate the token query
to let the frontend know that there is no valid token
and hence no user is logged in.